requirements.txt
# Core Dependencies
biopython==1.81
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Deep Learning
tensorflow==2.13.0
scikit-learn==1.3.0
keras==2.13.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Bioinformatics Tools
dendropy==4.6.1
ete3==3.1.3

# Configuration & Utilities
pyyaml==6.0.1
click==8.1.6
tqdm==4.65.0
joblib==1.3.1

# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.2